+++ src/ska_src_site_capabilities_api/rest/routers/nodes.py
@@ -11,7 +11,13 @@ from starlette.requests import Request
 from starlette.responses import HTMLResponse, JSONResponse
 
 from ska_src_site_capabilities_api import models
-from ska_src_site_capabilities_api.common.exceptions import IncorrectNodeVersionType, NodeAlreadyExists, SiteNotFoundInNodeVersion, handle_exceptions
+from ska_src_site_capabilities_api.common.exceptions import (
+    IncorrectNodeVersionType,
+    NodeAlreadyExists,
+    NodeVersionNotFound,
+    SiteNotFoundInNodeVersion,
+    handle_exceptions,
+)
 from ska_src_site_capabilities_api.common.utility import recursive_autogen_id
 from ska_src_site_capabilities_api.rest.dependencies import Common, Permissions
 from ska_src_site_capabilities_api.rest.logger import logger
@@ -229,14 +235,15 @@ async def get_node_version(
     node_version: str = Query(default="latest", description="Version of node ({version}||latest"),
 ) -> JSONResponse:
     """Get a version of a node."""
-    with LogContext(resource_id=node_name, operation="get_node"):
-        logger.info(f"Retrieving node: {node_name}, version: {node_version}")
-        if node_version != "latest":
-            try:
-                int(node_version)
-            except ValueError:
-                raise IncorrectNodeVersionType
-        return JSONResponse(request.app.state.backend.get_node(node_name=node_name, node_version=node_version))
+    if node_version != "latest":
+        try:
+            int(node_version)
+        except ValueError:
+            raise IncorrectNodeVersionType
+    node = request.app.state.backend.get_node(node_name=node_name, node_version=node_version)
+    if not node:
+        raise NodeVersionNotFound(node_name=node_name, node_version=node_version)
+    return JSONResponse(node)
 
 
 @api_version(1)